--- iree/compiler/Dialect/Flow/Conversion/TensorToFlow/CMakeLists.txt
@@ -1,28 +1,0 @@
-################################################################################
-# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# iree/compiler/Dialect/Flow/Conversion/TensorToFlow/BUILD                     #
-#                                                                              #
-# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
-# CMake-only content.                                                          #
-#                                                                              #
-# To disable autogeneration for this file entirely, delete this header.        #
-################################################################################
-
-iree_add_all_subdirs()
-
-iree_cc_library(
-  NAME
-    TensorToFlow
-  HDRS
-    "ConvertTensorToFlow.h"
-  SRCS
-    "ConvertTensorToFlow.cpp"
-  DEPS
-    MLIRIR
-    MLIRStandard
-    MLIRTensor
-    iree::compiler::Dialect::Flow::IR
-  PUBLIC
-)
-
-### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
