+++ iree/compiler/Dialect/Flow/Transforms/BUILD
@@ -65,7 +65,6 @@ cc_library(
     ],
     deps = [
         ":PassesIncGen",
-        "//iree/compiler/Dialect/Flow/Conversion/TensorToFlow",
         "//iree/compiler/Dialect/Flow/IR",
         "//iree/compiler/Dialect/HAL/IR",
         "//iree/compiler/Dialect/IREE/IR",
