+++ iree/compiler/InputConversion/Common/Passes.h
@@ -27,6 +27,15 @@ void registerCommonConversionPassPipelines();
 //===----------------------------------------------------------------------===//
 
 std::unique_ptr<OperationPass<FuncOp>> createTopLevelSCFToCFGPass();
+std::unique_ptr<OperationPass<FuncOp>> createConvertUpstreamToIREE();
+
+//===----------------------------------------------------------------------===//
+// Patterns
+//===----------------------------------------------------------------------===//
+
+void populateConvertUpstreamToIREEPatterns(MLIRContext *context,
+                                           TypeConverter &typeConverter,
+                                           OwningRewritePatternList &patterns);
 
 //===----------------------------------------------------------------------===//
 // Register all Passes
