+++ iree/compiler/InputConversion/Common/Passes.td
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef IREE_COMPILER_INPUTCONVERSION_COMMON_PASSES
-#define IREE_COMPILER_INPUTCONVERSION_COMMON_PASSES
+#ifndef IREE_COMPILER_INPUTCONVERSION_TOSA_PASSES
+#define IREE_COMPILER_INPUTCONVERSION_TOSA_PASSES
 
 include "mlir/Pass/PassBase.td"
 
@@ -15,4 +15,10 @@ def TopLevelSCFToCFG :
   let constructor = "mlir::iree_compiler::createTopLevelSCFToCFGPass()";
 }
 
-#endif // IREE_COMPILER_INPUTCONVERSION_COMMON_PASSES
+def ConvertUpstreamToIREE :
+    Pass<"iree-convert-upstream-to-iree", "FuncOp"> {
+  let summary = "Catch-all pass to convert upstream MLIR ops that (for whatever reason) we prefer to be represented differently in IREE";
+  let constructor = "mlir::iree_compiler::createConvertUpstreamToIREE()";
+}
+
+#endif // IREE_COMPILER_INPUTCONVERSION_TOSA_PASSES
